third_party/WebKit/LayoutTests/fast/js/iterable-object-expected.txt

Test ES6 iterator works with DOM objects.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS internals.iterator is undefined.
value = 0
value = 1
value = 4
value = 9
value = 16
value = 0
value = 1
value = 4
value = 9
value = 16
key = 0
key = 1
key = 2
key = 3
key = 4
value = 0
value = 1
value = 4
value = 9
value = 16
entry = 0,0
entry = 1,1
entry = 2,4
entry = 3,9
entry = 4,16
PASS successfullyParsed is true

TEST COMPLETE
